ObjectDetection/requirements.txt
pillow
torch
numpy
streamlit
shapely
opencv-python